Extract Cloud Storage ops into host/storage.ts
Move storage.listAll/getMetadata/getBlob/putBytes/getBytes/deleteObject with
the storage lens resolver (lensStorage/ensureStorage) and the wire metadata
mapper (toSettableMetadata) into host/storage.ts behind isStorageOp/
handleStorageOp. handleOp routes to it. Bodies verbatim; orphaned pyric/storage
imports dropped from host.ts. No behavior change.
